feat(updateURL): remove trailing '?' if no params provided
Updated readme file to correct install instructions and enhanced `updateURL` function in utils.js for a cleaner URL if no parameters provided.

The installation instructions in README.md have been modified to point to the correct repository and install using `npm` instead of `yarn`. In utils.js, a conditional statement has been added to exclude the trailing '?' from the URL if no parameters are provided, for cleaner URL structure.