feat(build): build assets + init ddev for easier development
Add instructions to README.md and create ddev commands for installing and building assets. Improve URL generation in dist/index.js.

This commit makes development easier by automating asset installation and build. It also improves URL construction in the javascript file to avoid having a URL ending with "?" when there is no value for the key in the query string.